+++ tests/Niirrty/IO/Vfs/Tests/VfsHandlerTest.php
@@ -12,20 +12,20 @@ namespace Niirrty\IO\Vfs\Tests;
 
 
 use Niirrty\ArgumentException;
-use Niirrty\IO\Vfs\Handler;
+use Niirrty\IO\Vfs\VfsHandler;
 use PHPUnit\Framework\TestCase;
 
 
-class HandlerTest extends TestCase
+class VfsHandlerTest extends TestCase
 {
 
 
    function testInit()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
-      $this->assertInstanceOf( Handler::class, $h );
-      $this->assertSame( 'Foo Handler', $h->getName() );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
+      $this->assertInstanceOf( VfsHandler::class, $h );
+      $this->assertSame( 'Foo VfsHandler', $h->getName() );
       $this->assertSame( 'foo://', $h->getProtocol() );
       $this->assertSame( 'foo', $h->getProtocolName() );
       $this->assertSame( '://', $h->getProtocolSeparator() );
@@ -38,16 +38,16 @@ class HandlerTest extends TestCase
    function testSetProtocolName()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__ );
-      $this->assertInstanceOf( Handler::class, $h->setProtocolName( 'bar' ) );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__ );
+      $this->assertInstanceOf( VfsHandler::class, $h->setProtocolName( 'bar' ) );
       $this->assertSame( 'bar', $h->getProtocolName() );
 
    }
    function testSetProtocolSeparator()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__ );
-      $this->assertInstanceOf( Handler::class, $h->setProtocolSeparator( ':/' ) );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__ );
+      $this->assertInstanceOf( VfsHandler::class, $h->setProtocolSeparator( ':/' ) );
       $this->assertSame( ':/', $h->getProtocolSeparator() );
 
    }
@@ -55,23 +55,23 @@ class HandlerTest extends TestCase
    {
 
       $this->expectException( ArgumentException::class );
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__ );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__ );
       $h->setRootFolder( __DIR__ . '/foobarbaz' );
 
    }
    function testIsValid()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__ );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__ );
       $this->assertTrue( $h->isValid() );
-      $h = new Handler( 'Foo Handler', '', '://', __DIR__ );
+      $h = new VfsHandler( 'Foo VfsHandler', '', '://', __DIR__ );
       $this->assertFalse( $h->isValid() );
 
    }
    function testAddReplacement()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__ );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__ );
       $this->assertFalse( $h->hasReplacement( 'blub' ) );
       $this->assertSame( $h, $h->addReplacement( 'blub', '1234' ) );
       $this->assertTrue( $h->hasReplacement( 'blub' ) );
@@ -82,7 +82,7 @@ class HandlerTest extends TestCase
    function testAddReplacements()
    {
 
-      $h = new Handler( 'Foo Handler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
+      $h = new VfsHandler( 'Foo VfsHandler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
       $this->assertTrue( $h->hasReplacement( 'blub' ) );
       $h->addReplacements( [ 'blub' => null, 'blubber' => '1212' ] );
       $this->assertFalse( $h->hasReplacement( 'blub' ) );
@@ -92,7 +92,7 @@ class HandlerTest extends TestCase
    function testTryParse()
    {
 
-      $h = Handler::Create( 'Foo Handler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
+      $h = VfsHandler::Create( 'Foo VfsHandler', 'foo', '://', __DIR__, [ 'blub' => 14 ] );
       $path = 'foo://bar/baz/${xyz}/${abc}';
       $this->assertTrue( $h->tryParse( $path, [ 'xyz' => '123' ] ) );
       $this->assertSame( __DIR__ . '/bar/baz/123/${abc}', $path );